- more improvements

+++ src/persist/services.ts
@@ -1,27 +1,30 @@
 import type { Request } from "express";
 import { Controller, Middleware, Route } from "../http/decorators";
 import { Inject, Service } from "../library";
-import type { EntityManagerInterface, GetManyResults } from "./decorators";
+import { PersistInterfaces, type EntityCollectionManagerInterface, type GetManyResults } from "./decorators";
 import { getModelDefinition } from "../entity/services";
 import { HttpError } from "../http/services";
 import { EntityValidationError } from "../entity/decorators";
 import { AppWeaverError } from "../constants";
 
 @Controller({ rootPath: '/_dangerous_override', isSubApp: true })
-@Service({ id: 'entityManagerLCRUDController.abstract' })
-export class EntityManagerLCRUDController {
+@Service({ id: 'persist.CollectionManagerLCRUDController.abstract', 
+    interfaces: [PersistInterfaces.EntityCollectionManager],
+ })
+export class CollectionManagerLCRUDController {
     @Inject('todo-entityManager-reference')
-    entityManager: EntityManagerInterface = undefined as unknown as EntityManagerInterface;
+    entityManager: EntityCollectionManagerInterface = undefined as unknown as EntityCollectionManagerInterface;
 
-    // @todo: add middleware for authentication/authorization
-    // @todo: add request hook e.g. normalizeEntity(modelName, entity, request): typeof entity
+    getRequestUserData(request: any) {
+        return {} as Record<string, any>;
+    }
 
     @Route({ path: '/:modelName/{:id}', methods: ['GET'] })
     async doGetOne(request: any, response: any) {
         const { modelName, id } = request.params;
-        console.log('doGetOne', {modelName, id, query: request.query });
+        const _ownerId = this.getRequestUserData(request).userId ?? 'anonymous';
         try {
-            return response.json(await this.entityManager.getOne(modelName as string, id));
+            return response.json(await this.entityManager.getOne(modelName as string, id, { _ownerId }));
         } catch (error) {
             return response.status(500).json({ error, modelName, id });
         }
@@ -30,9 +33,9 @@ export class EntityManagerLCRUDController {
     @Route({ path: '/:modelName', methods: ['GET'] })
     async doGetMany(request: Request, response: any) {
         const { modelName } = request.params;
-        console.log('doGetMany', {modelName, query: request.query });
+        const _ownerId = this.getRequestUserData(request).userId ?? 'anonymous';
         try {
-            return response.json(await this.entityManager.getMany(modelName as string, request.query as Record<string, any>));
+            return response.json(await this.entityManager.getMany(modelName as string, {...request.query as Record<string, any>, _ownerId}));
         } catch (error) {
             return response.status(500).json({ error, modelName });
         }
@@ -41,35 +44,36 @@ export class EntityManagerLCRUDController {
     @Route({ path: '/:modelName', methods: ['POST'], priority: 50 })
     async doCreate(request: any, response: any) {
         const { modelName } = request.params;
-        console.log('doCreate', {modelName, query: request.query, body: request.body });
+        const _ownerId = this.getRequestUserData(request).userId ?? 'anonymous';
         try {
-            return response.json(await this.entityManager.create(modelName, request.body));
+            return response.json(await this.entityManager.create(modelName, request.body, { _ownerId }));
         } catch (error) {
-            return EntityManagerLCRUDController.returnErrorResponse(error, request, response, modelName);
+            return CollectionManagerLCRUDController.returnErrorResponse(error, request, response, modelName);
         }
     }
 
     @Route({ path: '/:modelName/:id', methods: ['PUT'] })
     async doUpdate(request: any, response: any) {
         const { modelName, id: _id } = request.params;
+        const _ownerId = this.getRequestUserData(request).userId ?? 'anonymous';
         try {
             const data = request.body;
             // @todo inject filter from request query
-            const entity = await this.entityManager.getOne(modelName, _id);
+            const entity = await this.entityManager.getOne(modelName, _id, { _ownerId });
             if (!entity) { throw HttpError.notFound('entity-not-found', { path: request.path, modelName, id: _id }); }
             
             return response.json(await this.entityManager.update(modelName, {...data, _id}, {_id}));
         } catch (error) {
-            return EntityManagerLCRUDController.returnErrorResponse(error, request, response, modelName, _id);
+            return CollectionManagerLCRUDController.returnErrorResponse(error, request, response, modelName, _id);
         }
     }
 
     @Route({ path: '/:modelName/{:id}', methods: ['DELETE'] })
     async doDeleteOne(request: any, response: any) {
         const { modelName, id } = request.params;
-        console.log('doDeleteOne', {modelName, id, query: request.query });
+        const _ownerId = this.getRequestUserData(request).userId ?? 'anonymous';
         try {
-            const result = await this.entityManager.delete(modelName, id);
+            const result = await this.entityManager.delete(modelName, id, { _ownerId });
             if (result.deletedCount === 0) { throw HttpError.notFound('entity-not-found', { path: request.path, modelName, id }); }
             return response.json(result);
         } catch (error) {
